requirements.txt
torch
torchvision
Pillow
tk
